+++ src/main/java/com/viglet/shiohara/utils/ShObjectUtils.java
@@ -1,5 +1,6 @@
 package com.viglet.shiohara.utils;
 
+import java.util.List;
 
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.stereotype.Component;
@@ -7,16 +8,38 @@ import org.springframework.stereotype.Component;
 import com.viglet.shiohara.persistence.model.folder.ShFolder;
 import com.viglet.shiohara.persistence.model.object.ShObject;
 import com.viglet.shiohara.persistence.model.post.ShPost;
+import com.viglet.shiohara.persistence.model.site.ShSite;
 import com.viglet.shiohara.persistence.repository.object.ShObjectRepository;
 
 @Component
 public class ShObjectUtils {
 	@Autowired
-	private ShFolderUtils shFolderUtils;	
+	private ShFolderUtils shFolderUtils;
 	@Autowired
 	private ShObjectRepository shObjectRepository;
 	@Autowired
 	private ShPostUtils shPostUtils;
+
+	public boolean isVisiblePage(ShObject shObject) {
+		ShFolder shFolder = null;
+		if (shObject instanceof ShFolder) {
+			shFolder = (ShFolder) shObject;
+		} else if (shObject instanceof ShPost) {
+			ShPost shPost = (ShPost) shObject;
+			shFolder = shPost.getShFolder();
+		}
+		if (shFolder != null) {
+			List<ShFolder> breadcrumb = shFolderUtils.breadcrumb(shFolder);
+			if (breadcrumb.get(0).getName().equals("Home")) {
+				return true;
+			} else {
+				return false;
+			}
+		} else {
+			return false;
+		}
+
+	}
 
 	public String generateObjectLinkById(String objectId) {
 		if (objectId != null) {
@@ -34,5 +57,19 @@ public class ShObjectUtils {
 			}
 		}
 		return null;
+	}
+
+	public String generateObjectLink(ShObject shObject) {
+		return this.generateObjectLinkById(shObject.getId());
+	}
+
+	public ShSite getSite(ShObject shObject) {
+		if (shObject instanceof ShPost) {
+			return shPostUtils.getSite((ShPost) shObject);
+		} else if (shObject instanceof ShFolder) {
+			return shFolderUtils.getSite((ShFolder) shObject);
+		} else {
+			return null;
+		}
 	}
 }